requirements.txt
Flask==1.1.1
joblib==0.13.2
mecab-python3==0.996.2
numpy==1.17.0
pandas==0.25.0
scikit-learn==0.21.3
